NDRS-833: Extracat `LinearChainMetrics` to a file.
